Ferdigstiller tidligere forhåndsvarsel
